Disable feature pourcentage_of_correct_prediction
This feature is very long to complete (4s for long text)
