+++ detect_ai_content/ml_logic/for_texts/using_ml_features/TrueNetTextDecisionTreeClassifier.py
@@ -1,15 +1,13 @@
 
 from sklearn.preprocessing import RobustScaler
 from sklearn.model_selection import train_test_split, GridSearchCV
-from sklearn.pipeline import make_pipeline
 from sklearn.tree import DecisionTreeClassifier
-from sklearn.pipeline import make_pipeline, Pipeline
+from sklearn.pipeline import Pipeline
 
-from detect_ai_content.ml_logic.preprocess import preprocess
 from detect_ai_content.ml_logic.data import get_enriched_df
 from detect_ai_content.ml_logic.evaluation import evaluate_model
 from detect_ai_content.ml_logic.mlflow import mlflow_save_metrics, mlflow_save_model, mlflow_save_params, load_model
-from detect_ai_content.ml_logic.preprocess import preprocess, smartCleanerTransformer, smartEnrichTransformer, smartSelectionTransformer
+from detect_ai_content.ml_logic.preprocess import smartCleanerTransformer, smartEnrichTransformer, smartSelectionTransformer, smartEnrichFunction
 
 import os
 import pickle
@@ -49,7 +47,7 @@ class TrueNetTextDecisionTreeClassifier:
         }
 
         df = get_enriched_df()
-        X = preprocess(data=df, auto_enrich=False)
+        X = smartEnrichFunction(data=df)
         y = df['generated']
         grid = GridSearchCV(DecisionTreeClassifier(),param_grid,refit = True, verbose=2)
         grid.fit(X,y)
@@ -127,7 +125,6 @@ class TrueNetTextDecisionTreeClassifier:
             'text_corrections_ratio',
             'average_sentence_lenght',
             'average_neg_sentiment_polarity',
-            'pourcentage_of_correct_prediction',
             'lexical_diversity',
             'smog_index',
             'flesch_reading_ease',
