+++ detect_ai_content/ml_logic/for_texts/using_ml_features/TrueNetTextLogisticRegression.py
@@ -13,7 +13,7 @@ from detect_ai_content.ml_logic.mlflow import mlflow_save_metrics, mlflow_save_m
 from detect_ai_content.params import *
 from detect_ai_content.ml_logic.data import get_enriched_df
 from detect_ai_content.ml_logic.evaluation import evaluate_model
-from detect_ai_content.ml_logic.preprocess import preprocess, smartCleanerTransformer, smartEnrichTransformer, smartSelectionTransformer
+from detect_ai_content.ml_logic.preprocess import smartCleanerTransformer, smartEnrichTransformer, smartSelectionTransformer, smartEnrichFunction
 
 class TrueNetTextLogisticRegression:
     def local_trained_pipeline(self):
@@ -50,7 +50,7 @@ class TrueNetTextLogisticRegression:
         mlflow.start_run(experiment_id=experiment_id)
 
         df = get_enriched_df()
-        X = preprocess(data=df, auto_enrich=False)
+        X = smartEnrichFunction(data=df, auto_enrich=False)
         y = df['generated']
 
         model = LogisticRegression()
@@ -101,7 +101,6 @@ class TrueNetTextLogisticRegression:
             'text_corrections_ratio',
             'average_sentence_lenght',
             'average_neg_sentiment_polarity',
-            'pourcentage_of_correct_prediction',
             'lexical_diversity',
             'smog_index',
             'flesch_reading_ease',
